test: address final suite review findings
