+++ lib/src/features/stories/providers/story_manager_provider.dart
@@ -1,6 +1,7 @@
 import 'package:poptart/poptart.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:spark/src/core/network/atproto/data/models/feed_models.dart';
+import 'package:spark/src/core/network/atproto/data/repositories/story_repository.dart';
 import 'package:spark/src/core/utils/logging/logger.dart';
 import 'package:spark/src/features/stories/providers/story_auto_delete_provider.dart';
 import 'package:spark/src/features/stories/providers/story_provider_dependencies.dart';
@@ -33,25 +34,25 @@ class StoryManagerState {
 
 @riverpod
 class StoryManager extends _$StoryManager {
-  late final StoryProviderDependencies _dependencies;
+  late final StoryRepository _repository;
   late final SparkLogger _logger;
 
   @override
   Future<StoryManagerState> build() async {
-    _dependencies = ref.read(storyProviderDependenciesProvider);
-    _logger = _dependencies.loggerFor('StoryManager');
+    _repository = ref.read(storyRepositoryProvider);
+    _logger = ref.read(storyLoggerProvider('StoryManager'));
     ref.read(storyAutoDeleteExecutorProvider.future).ignore();
     return _loadInitial();
   }
 
   Future<StoryManagerState> _loadInitial() async {
     try {
-      final did = _dependencies.did;
+      final did = ref.read(storyCurrentDidProvider);
       if (did == null) {
         return StoryManagerState(stories: const [], error: 'Not authenticated');
       }
       // Page through all story records directly via atproto to include expired
-      if (!_dependencies.atprotoAvailable) {
+      if (!ref.read(storyAtprotoAvailableProvider)) {
         return StoryManagerState(
           stories: const [],
           error: 'AtProto not initialized',
@@ -60,7 +61,7 @@ class StoryManager extends _$StoryManager {
       String? cursor;
       final uris = <AtUri>[];
       do {
-        final result = await _dependencies.loadRecordPage(
+        final result = await _repository.listStoryRecords(
           did: did,
           cursor: cursor,
         );
@@ -72,9 +73,7 @@ class StoryManager extends _$StoryManager {
       if (uris.isEmpty) {
         return StoryManagerState(stories: const []);
       }
-      final storyViews = await _dependencies.storyRepository.getStoryViews(
-        uris,
-      );
+      final storyViews = await _repository.getStoryViews(uris);
 
       storyViews.sort((a, b) => b.indexedAt.compareTo(a.indexedAt));
 
@@ -98,7 +97,7 @@ class StoryManager extends _$StoryManager {
       final updatedList = List<StoryView>.from(current.stories)
         ..removeWhere((s) => s.uri == story.uri);
       state = AsyncData(current.copyWith(stories: updatedList));
-      await _dependencies.deleteRecord(story.uri);
+      await _repository.deleteStoryRecord(story.uri);
     } catch (e, s) {
       _logger.e('Error deleting story', error: e, stackTrace: s);
       // Revert by refreshing fully
